services/event-simulator/requirements.txt
# Core dependencies
kafka-python==2.0.2

# Phase 2 Lambda dependencies (not needed for Phase 1)
# boto3==1.28.85